agent-appoiments/requirements.txt
openai==1.58.1
python-dotenv==1.0.1
flask==3.1.0
flask-cors==5.0.0
langchain==0.3.13
langchain-openai==0.2.14
langgraph==0.2.59
requests==2.32.3
